Build daemons using their makefiles
**Note**: The existing clientmount services on compute nodes will need to
be removed. See below.

The nnf-dm and client-mount daemons were being built manually in
nnf-deploy rather than making use of their Makefiles. This was causing
the versions to sit at 0.0.0 rather than the built in versioning.
provided by the Makefiles.

In order to support this change, the client-mount service had to be
renamed to clientmountd in dws. The binary built by the Makefile and the
RPMs is named clientmountd, so nnf-deply looks for an
clientmountd.service file - which wouldn't exist without the rename.

Any compute node running the old version will need to remove the service
before the renamed service will work properly.

At LLNL, this should not be an issue since they are already using the
clientmountd binary, but they are writing their own service file.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>