Add k8sHost + k8sPort to systems config
The texas system's compute nodes cannot hit the k8s cluster with the
public facing IP like our other systems.

This change adds additional configuration for k8s host and port that is
fed into the daemons' configuration so they are able to talk to the
cluster from the computes using a separate host/port, if necessary.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>