Remove --nnf-node-name and --node-name from service files
Both nnf-dm and clientmountd default to the system hostname if these
options are not provided, so there is no need to keep providing them via
the systemd service file. LLNL does not plan on setting these command
line options to make configuration easier on their end - we should
follow suit.

The formatting for the overrides file has been cleaned up a bit as well.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>