Auto-create release on tag, with manifests
Releases are created when a v* tag is pushed to the releases/v0 branch.

As part of the release, the manifests from each of the submodules, and also
from cert-manager and mpi-operator, are collected into a tarball and included
as a release artifact.  This tarball is meant to be unpacked in a gitops repo.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>