Use kubectl create for SystemConfiguration resource
Use 'kubectl create' for this resource so that k8s does not create the
"last-applied-configuration" annotation.  This resource is not being updated by any controller, and the admins can be taught to update it with 'kubectl
replace' so that annotation is never needed.

This resource can be quite large on a big system and the
"last-applied-configuration" annotation that is created by 'kubectl apply' can
exceed the allowed size of an annotation.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>