Preserve the service namespace during undeploy
If a pod is on a host that cannot be reached, then any attempt to delete the
namespace will hang.  So change this to preserve the namespace and leave that
pod for k8s to cleanup when it can.

Add getopts to collect-manifests.sh, and update the tool to filter the
namespace out of the manifests it collects for ArgoCD. Make the tool all-in-one
so it collects the manifests and also makes the tarball; that should make it
less cumbersome in a manual situation.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>