kind.sh: fix reset
The script was calling itself and if you did this in the root directory
(now that `kind.sh` lives in `tools/`), it can't find `kind.sh`.

This just uses functions instead. There's still an issue with calling
`./nnf-deploy init` if you `cd tools/` first, but I'm ignoring that for
now.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>