Turn on ArgoCD via helm for nnf-deploy init
Use helm to install ArgoCD.  Do not install cert-manager, mpi-operator,
lustre-csi-driver, or lustre-fs-operator because those should be handled by
ArgoCD.

Create an overlay file for the config/repositories.yaml file, to revert it
back to its legacy pre-helm, pre-argocd behavior if desired.  To enable the
overlay, it must be copied into the root of the repo and named
'overlay-legacy.yaml'.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>