Collect separate manifests for hardware vs KIND
When collecting a manifest, specify whether it is being collected for KIND
or for hardware, and run the edit-image makefile target in each submodule
to ensure that the desired type of manifest will be generated. Read the
overlays list out of config/systems.yaml to find the appropriate overlay in
each submodule for the desired type of manifest.

Add both manifests to the release artifacts.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>