Ignore old versions when vendoring new API
Change the vendor-new-api.py script to only vendor a new API for the latest API
in the target repo. Old API versions may have embedded structs from a vendor API,
and we don't want them to change.

Also, add an option to specify which version of the vendor API to "go get".

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>