Allow vendoring into nnf-integration-test repo
Adjust vendor-new-api.py so it can vendor into the nnf-integration-test repo.
This relies on a new crd-bumper.yaml file in that repo to give some hints.

Refactor code that checks whether an API is a hub or a spoke.

Refactor code that references the crd-bumper.yaml file. This simplifies the
use of the object that holds the content of that file.

Reformat code with 'black'.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>