Let crd-bumper identify carry-over code for conversions
Define a new code marker to place around conversion code that must be
carried over to a new spoke. This should be used infrequently--it's almost
always wrong to carry over conversion code to a newly-created spoke.

This does not yet copy the marked code over to the new spoke, but it does
allow that code to be marked and it does find it and print out a banner about
it.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>