Mark an unserved API as served
It's possible that after marking an API as unserved and doing a release, there
will be a need to follow-up that release with a variation of it that does serve
that API. And maybe the commit that did the "unserve" step is buried under
other things and is not easily reverted. In that case, use the `re-serve.py`
tool.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>